requirements.txt
Flask==3.1.0
Flask-Mail==0.10.0
Flask-Limiter==3.9.2
python-dotenv==1.0.1